// src/api/generated/model/sessionCollectionSuggestionItem.ts
/**
 * Generated by orval v7.3.0 🍺
 * Do not edit manually.
 * Ludora API v1
 * API for the Ludora app
 * OpenAPI spec version: 1.0
 */
import type { SessionCollectionSuggestionItemGameMode } from './sessionCollectionSuggestionItemGameMode';
import type { TeamFromSessionCollectionItem } from './teamFromSessionCollectionItem';
import type { SessionCollectionSuggestionItemSport } from './sessionCollectionSuggestionItemSport';

export interface SessionCollectionSuggestionItem {
  /** Creator UID */
  readonly creatorUid: string;
  /** End date of the session */
  readonly endDate: string;
  /** Session field image */
  readonly fieldImage?: string;
  /** Field latitude */
  readonly fieldLatitude: number;
  /** Field longitude */
  readonly fieldLongitude: number;
  /** Session field short address */
  readonly fieldShortAddress: string;
  /** Game mode */
  readonly gameMode: SessionCollectionSuggestionItemGameMode;
  /** Maximum players per team */
  readonly maxPlayersPerTeam: number;
  /** Session teams */
  readonly sessionTeams: readonly TeamFromSessionCollectionItem[];
  /** Sport type */
  readonly sport: SessionCollectionSuggestionItemSport;
  /** Start date of the session */
  readonly startDate: string;
  /** Session unique identifier */
  readonly uid: string;
  /** Distance to the session in meters */
  readonly userDistance: number;
}